Replace pointer-based exec.rs with safe index-based implementation
Delete the old unsafe exec.rs and decode/primitives.rs. The safe
sequence execution loop (formerly safe_exec.rs) becomes the only
exec.rs, used by both default and paranoid builds.

On AVX2/NEON hardware the SIMD decode path handles all sequence
execution, so exec.rs is only the scalar fallback. Benchmarks
confirm zero measurable impact on the default build.
